+++ R/credentials_service_account.R
@@ -27,14 +27,15 @@ credentials_service_account <- function(scopes = NULL,
                                         path = "",
                                         ...,
                                         subject = NULL) {
-  ui_line("trying credentials_service_account()")
+  gargle_debug("trying {.fun credentials_service_account}")
   info <- jsonlite::fromJSON(path, simplifyVector = FALSE)
   if (!identical(info[["type"]], "service_account")) {
-    ui_line(
-      "JSON does not appear to represent a service account\n",
-      "did you provide the JSON for an OAuth client instead of for a ",
-      "service account?"
-    )
+    gargle_debug(c(
+      "JSON does not appear to represent a service account",
+      glue("
+        Did you provide the JSON for an OAuth client instead of for a \\
+        service account?")
+    ))
     return()
   }
 
@@ -53,7 +54,7 @@ credentials_service_account <- function(scopes = NULL,
     !nzchar(token$credentials$access_token)) {
     NULL
   } else {
-    ui_line("service account email: ", token_email(token))
+    gargle_debug("service account email: {.email {token_email(token)}}")
     token
   }
 }