+++ R/utils.R
@@ -28,7 +28,7 @@ is_rstudio_server <- function() {
   }
 }
 
-add_line <- function(path, line, quiet = FALSE) {
+add_line <- function(path, line) {
   if (file_exists(path)) {
     lines <- readLines(path, warn = FALSE)
     lines <- lines[lines != ""]
@@ -36,9 +36,11 @@ add_line <- function(path, line, quiet = FALSE) {
     lines <- character()
   }
 
-  if (line %in% lines) return(TRUE)
-  ui_line("Adding ", line, " to ", path)
+  if (line %in% lines) {
+    return(TRUE)
+  }
 
+  gargle_info("Adding {.val {line}} to {.file {path}}")
   lines <- c(lines, line)
   writeLines(lines, path)
   TRUE
@@ -55,7 +57,7 @@ normalize_scopes <- function(x) {
 }
 
 add_email_scope <- function(scopes = NULL) {
-  ui_line("adding 'userinfo.email' scope")
+  gargle_debug("adding {.val userinfo.email} scope")
   url <- "https://www.googleapis.com/auth/userinfo.email"
   union(scopes %||% character(), url)
 }