+++ tests/testthat/test-request_retry.R
@@ -1,4 +1,7 @@
 test_that("request_retry() logic works as advertised", {
+  # TODO: I'm testing too much re: retry logic via messages.
+  # Classed errors should simplify things, in due course.
+
   faux_response <- function(status_code = 200, h = NULL) {
     structure(
       list(status_code = status_code,
@@ -6,6 +9,7 @@ test_that("request_retry() logic works as advertised", {
       class = "response"
     )
   }
+  # allows us to replay a fixed set of responses
   faux_request_make <- function(responses = list(faux_response())) {
     i <- 0
     force(responses)
@@ -14,7 +18,35 @@ test_that("request_retry() logic works as advertised", {
       responses[[i]]
     }
   }
-  withr::local_options(list(gargle_quiet = FALSE))
+
+  # turn this: Retry 1 happens in 1.7 seconds
+  #            Retry 1 happens in 1 seconds
+  # into this: Retry 1 happens in {WAIT_TIME} seconds
+  fix_wait_time <- function(x) {
+    sub(
+      "(?<=in )[[:digit:]]+([.][[:digit:]]+)?(?= seconds)",
+      "{WAIT_TIME}",
+      x, perl = TRUE
+    )
+  }
+  # turn this: (strategy: exponential backoff, full jitter, clipped to floor of 1 seconds)
+  #            (strategy: exponential backoff, full jitter, clipped to ceiling of 45 seconds)
+  # into this: (strategy: exponential backoff, full jitter)
+  fix_strategy <- function(x) {
+    sub(
+      ", clipped to (floor|ceiling) of [[:digit:]]+([.][[:digit:]]+)? seconds",
+      "",
+      x, perl = TRUE
+    )
+  }
+  local_gargle_verbosity("debug")
+
+  # 2021-03-18
+  # The switch to mockr::with_mock() has caused trouble interactively
+  # executing this test. The closures used to mock request_make() seem to
+  # use a common counter (i) and responses. I plan to open a mockr issue.
+  # Workaround in the meantime: load_all() before each interactive call to
+  # with_mock().
 
   # succeed on first try
   out <- with_mock(
@@ -26,14 +58,18 @@ test_that("request_retry() logic works as advertised", {
 
   # fail, then succeed (exponential backoff)
   r <- list(faux_response(429), faux_response())
-  out <- with_mock(
+  with_mock(
     request_make = faux_request_make(r),
     gargle_error_message = function(...) "oops", {
-      request_retry(max_total_wait_time_in_seconds = 5)
+      msg_fail_once <- capture.output(
+        out <- request_retry(max_total_wait_time_in_seconds = 5),
+        type = "message"
+      )
     }
-  ) %>%
-    expect_message("Retry 1.*jitter") %>%
-    suppressMessages() # mops up messages that don't match regexp in line above
+  )
+  expect_snapshot(
+    writeLines(fix_strategy(fix_wait_time(msg_fail_once)))
+  )
   expect_equal(httr::status_code(out), 200)
 
   # fail, then succeed (Retry-After header)
@@ -41,14 +77,18 @@ test_that("request_retry() logic works as advertised", {
     faux_response(429, h = list(`Retry-After` = 1.4)),
     faux_response()
   )
-  out <- with_mock(
+  with_mock(
     request_make = faux_request_make(r),
     gargle_error_message = function(...) "oops", {
-      request_retry()
+      msg_retry_after <- capture.output(
+        out <- request_retry(),
+        type = "message"
+      )
     }
-  ) %>%
-    expect_message("Retry 1.*header") %>%
-    suppressMessages()
+  )
+  expect_snapshot(
+    writeLines(fix_strategy(fix_wait_time(msg_retry_after)))
+  )
   expect_equal(httr::status_code(out), 200)
 
   # make sure max_tries_total is adjustable)
@@ -58,14 +98,18 @@ test_that("request_retry() logic works as advertised", {
     faux_response(429),
     faux_response()
   )
-  out <- with_mock(
+  with_mock(
     request_make = faux_request_make(r[1:3]),
     gargle_error_message = function(...) "oops", {
-      request_retry(max_tries_total = 3, max_total_wait_time_in_seconds = 6)
+      msg_max_tries <- capture.output(
+        out <- request_retry(max_tries_total = 3, max_total_wait_time_in_seconds = 6),
+        type = "message"
+      )
     }
-  ) %>%
-    expect_message("Retry 2") %>%
-    suppressMessages()
+  )
+  expect_snapshot(
+    writeLines(fix_strategy(fix_wait_time(msg_max_tries)))
+  )
   expect_equal(httr::status_code(out), 429)
 })